Rename `UpdateEventFactory` to `UpdateEvents`.
