+++ testutil/src/main/java/org/spine3/examples/todolist/testdata/TestTaskEventFactory.java
@@ -216,9 +216,9 @@ public class TestTaskEventFactory {
         return result;
     }
 
-    public static class UpdateEventFactory {
+    public static class UpdateEvents {
 
-        private UpdateEventFactory() {
+        private UpdateEvents() {
         }
 
         /**
